pyproject.toml
[project]
name = "rag-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.7",
    "google-adk>=0.3.0",
    "langchain>=0.3.24",
    "marker-pdf>=1.6.2",
    "sentence-transformers>=4.1.0",
]
